+++ src/harlequin/cli.py
@@ -32,7 +32,7 @@ else:
 DEFAULT_ADAPTER = "duckdb"
 DEFAULT_LIMIT = 100_000
 DEFAULT_THEME = "harlequin"
-DEFAULT_KEYMAPS = ["vscode"]
+DEFAULT_KEYMAP_NAMES = ["vscode"]
 
 # configure the rich click interface (mostly --help options)
 DOCS_URL = "https://harlequin.sh/docs/getting-started"
@@ -76,6 +76,7 @@ click.rich_click.OPTION_GROUPS = {
                 "--show-files",
                 "--show-s3",
                 "--theme",
+                "--keymap-name",
                 "--limit",
                 "--config",
                 "--config-path",
@@ -211,15 +212,15 @@ def build_cli() -> click.Command:
         ),
     )
     @click.option(
-        "--keymap",
+        "--keymap-name",
         help=(
             "The name of a keymap plugin to load. Repeat this option to load "
             "multiple keymaps. Keymaps listed last will override earlier ones. "
-            "For example, to tweak the default keymap, use '--keymap vscode "
-            "--keymap my_keys'"
+            "For example, to tweak the default keymap, use '--keymap-name vscode "
+            "--keymap-name my_keys'"
         ),
         multiple=True,
-        default=["vscode"],
+        default=DEFAULT_KEYMAP_NAMES,
     )
     @click.option(
         "--config",
@@ -263,7 +264,9 @@ def build_cli() -> click.Command:
         """
         # load config from any config files
         try:
-            config = get_config_for_profile(config_path=config_path, profile=profile)
+            config, user_defined_keymaps = get_config_for_profile(
+                config_path=config_path, profile_name=profile
+            )
         except HarlequinConfigError as e:
             pretty_print_error(e)
             ctx.exit(2)
@@ -273,7 +276,7 @@ def build_cli() -> click.Command:
         for k in params:
             if (
                 ctx.get_parameter_source(k)
-                == click.core.ParameterSource.DEFAULT  # type: ignore
+                == click.core.ParameterSource.DEFAULT  # type: ignore[attr-defined]
             ):
                 kwargs.pop(k)
             # conn_str is an arg, not an option, so get_paramter_source is always CLI
@@ -281,7 +284,7 @@ def build_cli() -> click.Command:
                 kwargs.pop(k)
 
         # merge the config and the cli options
-        config.update(kwargs)
+        config.update(kwargs)  # type: ignore[typeddict-item]
 
         # detect and install (if necessary) a tzdatabase on Windows
         if sys.platform == "win32" and not config.pop("no_download_tzdata", None):
@@ -293,7 +296,7 @@ def build_cli() -> click.Command:
 
         # set the locale so we display numbers properly. Empty string uses system
         # default
-        locale_config: str = config.pop("locale", "")  # type: ignore
+        locale_config: str = config.pop("locale", "")
         try:
             set_locale(locale_config)
         except HarlequinLocaleError as e:
@@ -301,12 +304,13 @@ def build_cli() -> click.Command:
             ctx.exit(2)
 
         # remove the harlequin config from the options passed to the adapter
-        conn_str: Sequence[str] = config.pop("conn_str", tuple())  # type: ignore
+        conn_str: Sequence[str] = config.pop("conn_str", tuple())
         if isinstance(conn_str, str):
             conn_str = (conn_str,)
-        max_results: str | int = config.pop("limit", DEFAULT_LIMIT)  # type: ignore
-        theme: str = config.pop("theme", DEFAULT_THEME)  # type: ignore
-        show_files: Path | str | None = config.pop("show_files", None)  # type: ignore
+        max_results: str | int = config.pop("limit", DEFAULT_LIMIT)
+        theme: str = config.pop("theme", DEFAULT_THEME)
+        keymap_names: list[str] = config.pop("keymap_name", DEFAULT_KEYMAP_NAMES)
+        show_files: Path | str | None = config.pop("show_files", None)
         if show_files is not None:
             try:
                 show_files = Path(show_files)
@@ -315,23 +319,21 @@ def build_cli() -> click.Command:
                     HarlequinConfigError(msg=str(e), title="Harlequin Config Error")
                 )
                 ctx.exit(2)
-        show_s3: str | None = config.pop("show_s3", None)  # type: ignore
+        show_s3: str | None = config.pop("show_s3", None)
 
         # load and instantiate the adapter
-        adapter: str = config.pop("adapter", DEFAULT_ADAPTER)  # type: ignore
+        adapter: str = config.pop("adapter", DEFAULT_ADAPTER)
         adapter_cls: type[HarlequinAdapter] = adapters[adapter]
         try:
-            adapter_instance = adapter_cls(conn_str=conn_str, **config)
+            adapter_instance = adapter_cls(conn_str=conn_str, **config)  # type: ignore[misc]
         except HarlequinConfigError as e:
             pretty_print_error(e)
             ctx.exit(2)
 
-        # load specified keymaps
-        keymap_names: list[str] = config.pop("keymaps", DEFAULT_KEYMAPS)  # type: ignore
-
         tui = Harlequin(
             adapter=adapter_instance,
             keymap_names=keymap_names,
+            user_defined_keymaps=user_defined_keymaps,
             connection_hash=get_connection_hash(conn_str, config),
             max_results=max_results,
             theme=theme,
@@ -361,7 +363,7 @@ def build_cli() -> click.Command:
 
     fn = inner_cli
     for option in options.values():
-        fn = option.to_click()(fn)  # type: ignore
+        fn = option.to_click()(fn)  # type: ignore[assignment]
 
     return fn