+++ src/harlequin/config.py
@@ -2,11 +2,12 @@ from __future__ import annotations
 
 import sys
 from pathlib import Path
-from typing import Dict, List, Union
+from typing import Sequence, TypedDict
 
 from platformdirs import user_config_path
 
 from harlequin.exception import HarlequinConfigError
+from harlequin.keymap import HarlequinKeyMap, RawKeyBinding
 
 if sys.version_info < (3, 11):
     import tomli as tomllib
@@ -14,23 +15,50 @@ else:
     import tomllib
 
 
-Profile = Dict[str, Union[bool, int, List[str], str, Path]]
-Config = Dict[str, Union[str, Dict[str, Profile]]]
+class Profile(TypedDict, total=False):
+    conn_str: Sequence[str] | str
+    adapter: str
+    limit: str | int
+    theme: str
+    keymap_name: list[str]
+    show_files: Path | str | None
+    show_s3: str | None
+    locale: str
+    no_download_tzdata: bool
+    # many more keys for adapter options
 
 
-def get_config_for_profile(config_path: Path | None, profile: str | None) -> Profile:
+class Config(TypedDict, total=False):
+    default_profile: str | None
+    keymaps: dict[str, list[RawKeyBinding]]
+    profiles: dict[str, Profile]
+
+
+def get_config_for_profile(
+    config_path: Path | None, profile_name: str | None
+) -> tuple[Profile, list[HarlequinKeyMap]]:
     config = load_config(config_path)
-    active_profile = profile or config.get("default_profile", None)
-    if active_profile is None or active_profile == "None":
-        return {}
-    elif active_profile not in config.get("profiles", {}):
+    if not profile_name:
+        profile_name = config.get("default_profile", None)
+
+    if profile_name is None or profile_name == "None":
+        profile: Profile = {}
+    elif profile_name not in config.get("profiles", {}):
         raise HarlequinConfigError(
-            f"Could not load the profile named {active_profile} because it does not "
+            f"Could not load the profile named {profile_name} because it does not "
             "exist in any discovered config files.",
             title="Harlequin couldn't load your profile.",
         )
     else:
-        return config["profiles"][active_profile]  # type: ignore
+        profile = config["profiles"][profile_name]
+
+    raw_keymaps: dict[str, list[RawKeyBinding]] = config.get("keymaps", {})
+    keymaps: list[HarlequinKeyMap] = [
+        HarlequinKeyMap.from_config(name=name, bindings=bindings)
+        for name, bindings in raw_keymaps.items()
+    ]
+
+    return profile, keymaps
 
 
 def load_config(config_path: Path | None) -> Config:
@@ -138,7 +166,7 @@ def _merge_config_files(paths: list[Path]) -> Config:
 
 
 def _raise_on_bad_schema(config: Config) -> None:
-    TOP_LEVEL_KEYS = ("default_profile", "profiles")
+    TOP_LEVEL_KEYS = ("default_profile", "profiles", "keymaps")
     if not config:
         return
 
@@ -173,11 +201,33 @@ def _raise_on_bad_schema(config: Config) -> None:
             for option_name in opt_dict.keys():
                 if "-" in option_name:
                     raise HarlequinConfigError(
-                        f"Profile {profile_name} defines an option '{option_name}',"
+                        f"Profile {profile_name} defines an option {option_name!r}, "
                         "which is an invalid name for an option. Did you mean "
-                        f"""'{sluggify_option_name(option_name)}'?""",
+                        f"{sluggify_option_name(option_name)!r}?",
                         title="Harlequin couldn't load your config file.",
                     )
+                elif "keymap_names" in option_name:
+                    raise HarlequinConfigError(
+                        f"Profile {profile_name} defines an option {option_name!r}, "
+                        "which is an invalid name for an option. Did you mean "
+                        "'keymap_name' (singular)?",
+                        title="Harlequin couldn't load your config file.",
+                    )
+
+    if config.get("keymaps", None) is None:
+        pass
+    elif not isinstance(config["keymaps"], dict):
+        raise HarlequinConfigError(
+            "The keymaps key must define a table.",
+            title="Harlequin couldn't load your config file.",
+        )
+    elif not all(
+        [isinstance(config["keymaps"][k], list) for k in config["keymaps"].keys()]
+    ):
+        raise HarlequinConfigError(
+            "The members of each keymaps table must be arrays of tables.",
+            title="Harlequin couldn't load your config file.",
+        )
 
     if (default := config.get("default_profile", None)) is not None and not isinstance(
         default, str