+++ src/harlequin/plugins.py
@@ -1,9 +1,10 @@
 from __future__ import annotations
 
 import sys
-from typing import Literal, overload
+from typing import Literal, Sequence, overload
 
 from harlequin.adapter import HarlequinAdapter
+from harlequin.exception import HarlequinConfigError
 from harlequin.keymap import HarlequinKeyMap
 
 if sys.version_info < (3, 10):
@@ -16,8 +17,24 @@ def load_adapter_plugins() -> dict[str, type[HarlequinAdapter]]:
     return _load_plugins(group="harlequin.adapter")
 
 
-def load_keymap_plugins() -> dict[str, HarlequinKeyMap]:
-    return _load_plugins(group="harlequin.keymap")
+def load_keymap_plugins(
+    user_defined_keymaps: Sequence[HarlequinKeyMap],
+) -> dict[str, HarlequinKeyMap]:
+    keymaps = _load_plugins(group="harlequin.keymap")
+    for keymap in user_defined_keymaps:
+        if keymap.name in keymaps:
+            raise HarlequinConfigError(
+                title="Harlequin could not load your keymap config",
+                msg=(
+                    "Your Harlequin config files define a keymap named "
+                    f"{keymap.name}, but that name is already defined by "
+                    "a plug-in keymap. To extend a plug-in keymap, define "
+                    "a keymap with a new name, and configure your profile to "
+                    "load both keymaps."
+                ),
+            )
+        keymaps[keymap.name] = keymap
+    return keymaps
 
 
 @overload